Backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
lightgbm
joblib
tqdm
python-dotenv
